+++ src/pages/ShoppingCartPage/ShoppingCartPage.tsx
@@ -1,5 +1,3 @@
-import { useState } from 'react';
-import { useRecoilState } from 'recoil';
 import { useLoaderData, Await, Link } from 'react-router-dom';
 import Header, { MainLogoButton } from '../../components/Header/Header';
 import TitleContainer from '../../components/common/TitleContainer/TitleContainer';
@@ -7,43 +5,18 @@ import SubmitButton from '../../components/common/SubmitButton/SubmitButton';
 import CartItemList from '../../components/ShoppingCartPage/CartItemList/CartItemList';
 import TotalPriceContainer from '../../components/common/TotalPriceContainer/TotalPriceContainer';
 import type { TCartItem } from '../../types/CartItem.type';
-import { selectedCartItemListState } from '../../recoil/CartItem/atoms/atoms';
-import { fetchCartItemList } from '../../apis';
 import { EmptyCart } from '../../assets';
 import { PATHS } from '../../constants/PATHS';
 import * as S from './ShoppingCartPage.style';
+import useCartItems from '../../hooks/useCartItems';
 
 function ShoppingCartPage() {
-  const initialValue = useLoaderData() as TCartItem[];
+  const initialCartItemList = useLoaderData() as TCartItem[];
 
-  const [cartItemList, setCartItemList] = useState<TCartItem[]>(initialValue);
-
-  const [selectedItemList, setSelectedItemList] = useRecoilState(selectedCartItemListState);
-
-  const updateCartItemList = async () => {
-    const newCartItemList = await fetchCartItemList();
-    setCartItemList(newCartItemList);
-
-    const newSelectedItemList = newCartItemList.filter((el) => selectedItemList.some((item) => el.id === item.id));
-    setSelectedItemList(newSelectedItemList);
-  };
+  const { cartItemList, selectedItemList, updateCartItemList } = useCartItems(initialCartItemList);
 
   const hasCartItemList = cartItemList.length !== 0;
   const hasSelectedCartItemList = selectedItemList.length !== 0;
-
-  const EmptyCartContainer = (
-    <S.CartEmptyContainer>
-      <img src={EmptyCart} alt="빈 장바구니" />
-      <p>장바구니에 담은 상품이 없습니다.</p>
-    </S.CartEmptyContainer>
-  );
-
-  const CartListContainer = (
-    <>
-      <CartItemList cartItemList={cartItemList} updateCartItemList={updateCartItemList} />
-      <TotalPriceContainer />
-    </>
-  );
 
   return (
     <div>
@@ -56,7 +29,17 @@ function ShoppingCartPage() {
             title="장바구니"
             subTitle={cartItemList.length !== 0 ? `현재 ${cartItemList.length}종류의 상품이 담겨있습니다.` : ''}
           />
-          {hasCartItemList ? CartListContainer : EmptyCartContainer}
+          {hasCartItemList ? (
+            <>
+              <CartItemList cartItemList={cartItemList} updateCartItemList={updateCartItemList} />
+              <TotalPriceContainer />
+            </>
+          ) : (
+            <S.CartEmptyContainer>
+              <img src={EmptyCart} alt="빈 장바구니" />
+              <p>장바구니에 담은 상품이 없습니다.</p>
+            </S.CartEmptyContainer>
+          )}
         </S.Main>
       </Await>
       {hasSelectedCartItemList ? (
